Create hw01-q1.py
Task 1:
Write a program that will take three parameters of integer type from user and display the table of the
required value from the starting limit to ending limit (starting limit < ending limit). The first 
parameter will be the table value, second will be the starting limit and third will be the ending limit.